make clang-format check linux-only and adapt to trusty

+++ .ci/travis-compile.sh
@@ -22,26 +22,28 @@ if [[ $BUILDTYPE == "Debug" ]]; then
   make -j2
   make test
 
-  clang-format -i \
-    common/*.h \
-    common/*.cpp \
-    cockatrice/src/*.h \
-    cockatrice/src/*.cpp \
-    oracle/src/*.h \
-    oracle/src/*.cpp \
-    servatrice/src/*.h \
-    servatrice/src/*.cpp
-  
-  git clean -f
-  git diff --quiet || (
-    echo "*****************************************************";
-    echo "***  This PR is not clean against our code style  ***";
-    echo "***  Run clang-format and fix up any differences  ***";
-    echo "***  Check our CONTRIBUTING.md file for details!  ***";
-    echo "***                  Thank you ♥                  ***";
-    echo "*****************************************************";
-  )
-  git diff --exit-code
+  if [[ $TRAVIS_OS_NAME == "linux" ]]; then
+    clang-format-3.9 -i \
+      common/*.h \
+      common/*.cpp \
+      cockatrice/src/*.h \
+      cockatrice/src/*.cpp \
+      oracle/src/*.h \
+      oracle/src/*.cpp \
+      servatrice/src/*.h \
+      servatrice/src/*.cpp
+    
+    git clean -f
+    git diff --quiet || (
+      echo "*****************************************************";
+      echo "***  This PR is not clean against our code style  ***";
+      echo "***  Run clang-format and fix up any differences  ***";
+      echo "***  Check our CONTRIBUTING.md file for details!  ***";
+      echo "***                  Thank you ♥                  ***";
+      echo "*****************************************************";
+    )
+    git diff --exit-code
+  fi
 fi
 
 if [[ $BUILDTYPE == "Release" ]]; then
